+++ libraries/base/changelog.md
@@ -8,6 +8,8 @@
 
   * Add `Data.List.sortOn` sorting function
 
+  * Weaken RealFloat constraints on some `Data.Complex` functions
+
 ## 4.7.0.0  *Apr 2014*
 
   * Bundled with GHC 7.8.1